Wrote code for the hamming nums kata
Solved it but it times out on code wars. Have been trying to optimize it for the past however long but am nowhere near optimized enough. Guessing there is a mathematical formula to calculate this number easily. I simply do not know enough about these "hamming" numbers or clojure to try to implement/calculate a formula I do not know about. I will revisit this in the future.